ability to emojify text, and use emoji in llm toxicity detection
in the past you could bypass toxicity checks by replacing an offensive
word with an emoji. for eg: I think you are a <pig emoji>. In this pr we
convert all emojis to text before sending it to the local llm, so it
catches such scenarios. we also replace :emoji: name with the actual
emoji